update refine server API and refactor

+++ coffee_shop_backend/db_manager.py
@@ -28,7 +28,7 @@ class User(Base):
     def __init__(self, uid, password, type):
         self.uid = uid
         self.password = password
-        self.type=type
+        self.user_type=type
     
     def return_user_info(self):
         return {"uid":self.uid}
@@ -88,7 +88,20 @@ class PurchaseHistory(Base):
                 f"purchase_options={self.purchase_options}, price={self.price}, "
                 f"transaction_status={'Received' if self.transaction_status else 'Pending'}, "
                 f"cooking_process={self.cooking_process}, uid={self.uid})>")
+    
+class LoginSession(Base):
+    __tablename__ = 'login_session'
+    id = Column(String, primary_key=True, default=str(uuid.uuid4))
+    session_token = Column(String, nullable=False)
+    uid = Column(String, ForeignKey('Users.uid'), nullable=False)  # Foreign key to user
 
+    def __init__(self, session_token, uid):
+        self.session_token=session_token
+        self.uid=uid
+
+    def __repr__(self):
+        return f"<LoginSession(session_token={self.session_token})>"
+    
 class DBManager():
     def __init__(self, base, addr="sqlite:///serverdb.db") -> None:
         self.Base = base #declarative_base()
@@ -146,6 +159,44 @@ class DBManager():
 
         session.close()
         return users
+    
+    def find_user_by_uid_and_type(self, uid, user_type='customer'): #'admin'
+        
+        session = self.Session()
+        
+        user = session.query(User).filter((User.uid==uid) & (User.user_type==user_type)).first()
+
+        session.close()
+        return user
+    
+    def add_login_session(self, token,uid):
+        session = self.Session()
+        session.add(LoginSession(session_token=token,uid=uid))
+        session.commit()
+        session.close()
+    
+    def delete_login_session(self, token):
+        session = self.Session()
+        loginsession = session.query(LoginSession).filter(LoginSession.session_token==token).first()
+        if loginsession:
+            session.delete(loginsession)
+        session.commit()
+        session.close()
+    def delete_login_session_uid(self, uid):
+        session = self.Session()
+        loginsession = session.query(LoginSession).filter(LoginSession.uid==uid).first()
+        if loginsession:
+            session.delete(loginsession)
+        session.commit()
+        session.close()
+    def verify_login_session(self,token):
+        session = self.Session()
+        loginsession = session.query(LoginSession).filter(LoginSession.session_token==token).first()
+        session.close()
+        if loginsession:
+            return loginsession.uid
+        else:
+            return None
 
     def add_product(self, product_name, product_type, product_prices, product_options, product_status = True):
         # product_id = Column(Integer, primary_key=True, autoincrement=True)
@@ -182,8 +233,9 @@ class DBManager():
         return products
     def delete_product_by_name(self, name):
         session = self.Session()
-        product = session.query(Product).filter_by(name=name).first()
-        session.delete(product)
+        product = session.query(Product).filter(Product.product_name == name).first()
+        if product:
+            session.delete(product)
 
         session.commit()
         session.close()
@@ -234,9 +286,18 @@ class DBManager():
         session.query(PurchaseHistory).filter(PurchaseHistory.id==history_id).update({"cooking_process":cooking_process})
         session.commit()
         session.close()
+
+
 # DBM = DBManager(base=Base)
+# DBM.add_user(uid='abc@a.cn',password='123',type='customer')
+
+
+
 # session = DBM.Session()
 # users= session.query(User).all()
 # session.close()
+# print(users[0].user_type)
 
-# print(users)
+# uu=DBM.find_user_by_uid_and_type(uid='abc@a.cn', user_type='customer')
+# print(uu)
+
